moved smi package
