+++ smi/smi.go
@@ -1,4 +1,4 @@
-package main
+package smi
 
 import (
 	"fmt"
